python-app/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
jinja2==3.1.6
python-multipart==0.0.20
pydantic-settings==2.10.1
pypdf==5.9.0
openpyxl==3.1.5
certifi==2025.8.3
pyOpenSSL==26.0.0
pypyodbc==1.3.6
